server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
    #   django-stubs
black==25.1.0
    # via -r requirements.in
build==1.2.2.post1
    # via pip-tools
certifi==2025.1.31
    # via requests
chardet==5.2.0
    # via reportlab
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   black
    #   pip-tools
django==5.1.7
    # via
    #   django-cors-headers
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
django-cors-headers==4.7.0
    # via -r requirements.in
django-stubs==5.1.3
    # via djangorestframework-stubs
django-stubs-ext==5.1.3
    # via django-stubs
djangorestframework==3.15.2
    # via -r requirements.in
djangorestframework-stubs==3.15.3
    # via -r requirements.in
idna==3.10
    # via requests
isort==6.0.1
    # via -r requirements.in
mypy==1.15.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==24.2
    # via
    #   black
    #   build
pathspec==0.12.1
    # via black
pdf2image==1.17.0
    # via -r requirements.in
pillow==11.1.0
    # via
    #   -r requirements.in
    #   pdf2image
    #   reportlab
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.7
    # via black
polyline==2.0.2
    # via -r requirements.in
psycopg==3.2.6
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via -r requirements.in
reportlab==4.3.1
    # via -r requirements.in
requests==2.32.3
    # via djangorestframework-stubs
sqlparse==0.5.3
    # via django
types-pyyaml==6.0.12.20241230
    # via
    #   django-stubs
    #   djangorestframework-stubs
types-requests==2.32.0.20250306
    # via djangorestframework-stubs
typing-extensions==4.12.2
    # via
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
    #   mypy
    #   psycopg
urllib3==2.3.0
    # via
    #   requests
    #   types-requests
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
